Create AccordionList directory, move files and update import paths
